+++ src/frontend/apps/web/src/App.tsx
@@ -10,8 +10,21 @@ import {
 const dataContext = parseDataContext(document);
 
 // todo - brainstorm on an autodiscovery mechanism for routes
+const videoRoutes: Routes = {
+  instructor: lazy(() => import("@openfun/warren-video/src/pages/Instructor")),
+  student: lazy(() => import("@openfun/warren-video/src/pages/Student")),
+};
+
+const getVideoRoute = (): React.LazyExoticComponent<() => JSX.Element> => {
+  if (dataContext.is_instructor) {
+    return videoRoutes.instructor;
+  } else {
+    return videoRoutes.student;
+  }
+};
+
 const routes: Routes = {
-  demo: lazy(() => import("@openfun/warren-video/src/pages/Overview")),
+  video: getVideoRoute(),
 };
 
 export const App = () => {