+++ src/frontend/packages/video/src/pages/Student/index.tsx
@@ -1,0 +1,16 @@
+/**
+ * A React component responsible for rendering a dashboard of video statistics
+ * for student.
+ *
+ * This component combines the Filters component for selecting date ranges and videos, with the DailyViews component
+ * to display daily video statistics. It serves as a dashboard overview of all videos' statistical data.
+ *
+ * @returns {JSX.Element} The JSX for the videos statistics overview dashboard.
+ */
+export default () => {
+  return (
+    <div className="c__overview">
+      No student view available for video indicators
+    </div>
+  );
+};