+++ src/api/core/warren/__init__.py
@@ -1,3 +1,3 @@
 """Warren package."""
 
-__version__ = "0.2.0"
+__version__ = "0.3.0"